refactor: 패키지 수정
core, model 패키지 -> domain 패키지
