+++ src/test/kotlin/racingcar/CarTest.kt
@@ -2,7 +2,7 @@ package racingcar
 
 import org.junit.jupiter.api.Assertions.assertTrue
 import org.junit.jupiter.api.Test
-import racingcar.model.Car
+import racingcar.domain.Car
 
 class CarTest {
     @Test
